+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 See [RELEASE](./RELEASE.md) for workflow instructions.
 
+## UNRELEASED
+### Improvements
+* [#6035](https://github.com/spacemeshos/go-spacemesh/issues/6035) Fixed an issue where the node retried registering for the PoET round 
+  only for 15-20 minutes instead of continuing until the start of the round
+
 ## Release v1.6.1
 
 ### Improvements